add new prompts + versioning
