+++ app/lib/gas/v1/prompt/candidate/__init__.py
@@ -1,0 +1,7 @@
+from .base import prompt
+
+candidates = {
+    "gas_optimization_1": prompt,
+    "gas_optimization_2": prompt,
+    "gas_optimization_3": prompt,
+}
