+++ app/lib/security/v1/__init__.py
@@ -1,0 +1,7 @@
+from app.schema.llm import VersionDict
+
+from .markdown import template
+from .prompt import prompts
+from .response import OutputStructure
+
+structure = VersionDict(markdown=template, prompts=prompts, response=OutputStructure)
