+++ app/lib/security/v1/markdown/__init__.py
@@ -1,5 +1,5 @@
-markdown = """
-# ⛽ Smart Contract Gas Audit Report
+template = """
+# ⛽ Smart Contract Security Audit Report
 
 ---
 
@@ -9,17 +9,17 @@ Produced by: CertaiK AI Agent
 ---
 
 ⚠️ Severity Level Definitions
-- Critical: 🚨 Significant gas inefficiencies that can lead to substantial cost increases.
-- High: 🔴 Major gas usage issues that may result in noticeable cost impacts.
-- Medium: 🟠 Moderate gas optimization opportunities with potential cost savings.
-- Low: 🟢 Minor gas inefficiencies with limited cost impact.
+- Critical: 🚨 Issues that can lead to contract compromise or significant financial losses.  
+- High: 🔴 Severe bugs that may result in major exploits or disruptions.  
+- Medium: 🟠 Moderate risks with potential functional or security impacts.  
+- Low: 🟢 Minor issues with limited risk or impact.  
 
 ---
 
 ## 📝 Audit Summary
 - Contract Address: {address}
-- Audit Data: {date}
-- Auditors: Certaik AI Agent
+- Audit Date: {date}
+- Auditor: Certaik AI Agent
 
 ---
 
@@ -46,6 +46,7 @@ Produced by: CertaiK AI Agent
 
 ### 🟢 Low
 {findings_low}
+
 
 ---
