+++ app/lib/security/v1/prompt/__init__.py
@@ -1,0 +1,6 @@
+from app.schema.llm import Candidates
+
+from .candidate import candidates
+from .reviewer.base import prompt
+
+prompts = Candidates(candidates=candidates, reviewer=prompt)
