--- app/lib/v1/prompts/gas.py
@@ -1,135 +1,0 @@
-from typing import List
-
-from pydantic import BaseModel, Field
-
-candidate_prompt = """
-You are a professional smart contract auditor.
-Given smart contract code, you are tasked with identifying gas optimization opportunities.
-This does not have to be a formally structured report, but your answer should be complete.
-
-When producing your findings, be sure to include direct references to code including the line, variable, and function name, along with detailed descriptions of the vulnerability.
-
-Importantly, each gas optimization you find must fall into 1 of these classifications:
-- Critical: A gas inefficiency that can lead to substantial cost increases.
-- High: A major gas usage issue that may result in noticeable cost impacts.
-- Medium: A moderate gas optimization opportunity with potential cost savings.
-- Low: A minor gas inefficiency with limited cost impact.
-
-
-To further assist you, here are some basic principles for gas optimization that you should enforce when applicable:
-1. Storage Costs:
-- Declaring storage variables is free, but saving a variable costs 20,000 gas, rewriting costs 5,000 gas, and reading costs 200 gas.
-- Optimize by using memory for calculations before updating storage.
-
-2. Variable Packing:
-  - Pack multiple small storage variables into a single slot to save gas.
-  - Use \`bytes32\` for optimized storage and pack structs efficiently.
-
-3. Initialization:
-  - Avoid initializing zero values; default values are zero.
-
-4. Constants:
-  - Use \`constant\` for immutable values to save gas.
-
-5. Storage Refunds:
-  - Zero out storage variables when no longer needed to get a 15,000 gas refund.
-
-6. Data Types:
-  - Prefer \`bytes32\` over \`string\` for fixed-size data.
-  - Use fixed-size arrays and variables for efficiency.
-
-7. Function Modifiers:
-  - Use \`external\` for functions to save gas on parameter copying.
-  - Minimize public variables and use private visibility.
-
-8. Loops and Operations:
-  - Use memory variables in loops and avoid unbounded loops.
-  - Use \`++i\` instead of \`i++\` for gas efficiency.
-
-9. Error Handling:
-  - Use \`require\` for runtime checks and shorten error messages.
-
-10. Hash Functions:
-    - Prefer \`keccak256\` for hashing due to lower gas costs.
-
-11. Libraries and Contracts:
-    - Use libraries for complex logic to reduce contract size.
-    - Consider EIP1167 for deploying multiple contract instances.
-
-12. Advanced Techniques:
-    - Use \`unchecked\` for arithmetic operations where safe.
-    - Explore Yul for low-level optimizations.
-
-Do not make up findings. There are severe implications of the results you produce. It is okay if there are no gas optimization opportunities for a certain category.
-Please adhere strictly to the guidelines provided and don't make mistakes.
-
-The code to be audited for the gas optimization audit is found below:
-"""
-
-reviewer_prompt = """
-You are a smart contract gas optimization audit reviewer.
-You are given a smart contract and several professional reports, and are tasked with critiquing and prioritizing the gas optimizations that each auditor found.
-If a provided report includes findings that you believe are not real, make sure to state this.
-
-You are free to use your own discretion when determining the validity of these findings. You are a professional smart contract auditor, after all, but these findings should provide a strong basis of knowledge.
-
-Importantly, each gas optimization you find must fall into 1 of these classifications:
-- Critical: A gas inefficiency that can lead to substantial cost increases.
-- High: A major gas usage issue that may result in noticeable cost impacts.
-- Medium: A moderate gas optimization opportunity with potential cost savings.
-- Low: A minor gas inefficiency with limited cost impact.
-
-You should pay careful attention to deduplicating findings, and correctly classifying findings that more than 1 auditor might have found, but classified differently.
-
-Do not make up findings. There are severe implications of the results you produce. It is okay if there are no gas optimization findings for a certain category.
-
-Be certain to retain the direct references to code, if the auditors included it.
-"""
-
-reporter_prompt = """
-You are a smart contract gas optimization audit report generator.
-You are given the critique from an auditing professional, and are tasked with generating a structured report with severity classifications, explanations, and recommendations.
-
-You should carefully classify each vulnerability and strictly follow the expected output structure. You should not reference an auditor directly in your report.
-"""
-
-
-class FindingType(BaseModel):
-    name: str = Field(description="Name of the gas optimization finding")
-    explanation: str = Field(
-        description="Description of the gas optimization, with code references if applicable"
-    )
-    recommendation: str = Field(
-        description="Recommended action to take to resolve the gas optimization"
-    )
-    reference: str = Field(
-        description="A reference to the line of code and variable/function related to the gas optimization"
-    )
-
-
-class FindingsStructure(BaseModel):
-    critical: List[FindingType] = Field(
-        description="A list of critical gas optimizations, if any"
-    )
-    high: List[FindingType] = Field(
-        description="A list of high severity gas optimizations, if any"
-    )
-    medium: List[FindingType] = Field(
-        description="A list of medium severity gas optimizations, if any"
-    )
-    low: List[FindingType] = Field(
-        description="A list of low severity gas optimizations, if any"
-    )
-
-
-class OutputStructure(BaseModel):
-    introduction: str = Field(
-        description="a brief introduction of the smart contract's function"
-    )
-    scope: str = Field(
-        description="a brief overview of the scope of the smart contract audit"
-    )
-    findings: FindingsStructure = Field(
-        description="a detailed object of gas optimization findings"
-    )
-    conclusion: str = Field(description="a brief summary of the audit report")
